[src] Define torchhub utility in audyn/
