+++ audyn/utils/torchhub/__init__.py
@@ -1,0 +1,10 @@
+from .ast import ast_base
+from .passt import passt_base
+from .ssast import multitask_ssast_base_400, ssast_base_400
+
+__all__ = [
+    "ast_base",
+    "multitask_ssast_base_400",
+    "ssast_base_400",
+    "passt_base",
+]
